Revert "Audiocall code refactor"
