+++ src/components/AudiocallGroupItem/AudiocallGroupItem.tsx
@@ -1,6 +1,11 @@
-import { FC, PropsWithChildren } from 'react';
-import AudiocallGroupItemProps from '../../interfaces/audiocallGroupItemProps';
+import { FC, MouseEvent, PropsWithChildren } from 'react';
 import s from './AudiocallGroupItem.module.scss';
+
+interface AudiocallGroupItemProps {
+  active?: boolean;
+  className?: string | string[];
+  onClick?: (e: MouseEvent<HTMLButtonElement>) => void;
+}
 
 const AudiocallGroupItem: FC<PropsWithChildren<AudiocallGroupItemProps>> = ({
   active,
