+++ src/components/AudiocallGroupList/AudiocallGroupList.tsx
@@ -4,7 +4,10 @@ import wordsGroupNames from '../../shared/wordsGroupNames';
 import { useAppSelector } from '../../hooks/redux';
 import AudiocallGroupItem from '../AudiocallGroupItem';
 import { selectAudiocallGroup } from '../../store/audiocall/audiocallSlice';
-import AudiocallGroupListProps from '../../interfaces/audiocallGroupListProps';
+
+export interface AudiocallGroupListProps {
+  onClickItem: (groupName: string) => void;
+}
 
 const AudiocallGroupList: FC<AudiocallGroupListProps> = ({ onClickItem }) => {
   const activeGroup: number = useAppSelector(selectAudiocallGroup);
