--- src/interfaces/audiocallAnswerInfo.ts
@@ -1,6 +1,0 @@
-interface AudiocallAnswerInfo {
-  word: string;
-  wordIndex: number;
-}
-
-export default AudiocallAnswerInfo;
