--- src/interfaces/audiocallGroupItemProps.ts
@@ -1,9 +1,0 @@
-import { MouseEvent } from 'react';
-
-interface AudiocallGroupItemProps {
-  active?: boolean;
-  className?: string | string[];
-  onClick?: (e: MouseEvent<HTMLButtonElement>) => void;
-}
-
-export default AudiocallGroupItemProps;
