--- src/interfaces/audiocallGroupListProps.ts
@@ -1,5 +1,0 @@
-interface AudiocallGroupListProps {
-  onClickItem: (groupName: string) => void;
-}
-
-export default AudiocallGroupListProps;
