--- src/interfaces/audiocallMeaningData.ts
@@ -1,7 +1,0 @@
-interface AudiocallMeaningData {
-  imageLink: string;
-  imageAlt: string;
-  currentWord: string;
-}
-
-export default AudiocallMeaningData;
