--- src/interfaces/audiocallState.ts
@@ -1,13 +1,0 @@
-import { Word } from './words';
-
-interface AudiocallState {
-  audiocallGroup: number;
-  audiocallPage: number;
-  audiocallCurrentWord: number;
-  audiocallAnswers: { word: string; wordIndex: number }[];
-  audiocallShouldContinue: boolean;
-  audiocallWrongAnswers: Word[];
-  audiocallCorrectAnswers: Word[];
-}
-
-export default AudiocallState;
