--- src/shared/userWordDifficulty.ts
@@ -1,6 +1,0 @@
-enum UserWordDifficulty {
-  weak = 'weak',
-  hard = 'hard',
-}
-
-export default UserWordDifficulty;
