+++ src/store/audiocall/audiocallSlice.ts
@@ -1,15 +1,17 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 import type { RootState } from '..';
-import AudiocallAnswerInfo from '../../interfaces/audiocallAnswerInfo';
-import AudiocallState from '../../interfaces/audiocallState';
 import { Word } from '../../interfaces/words';
+
+interface AudiocallState {
+  audiocallGroup: number;
+  audiocallPage: number;
+  audiocallWrongAnswers: Word[];
+  audiocallCorrectAnswers: Word[];
+}
 
 const initialState: AudiocallState = {
   audiocallGroup: 0,
   audiocallPage: 0,
-  audiocallCurrentWord: 0,
-  audiocallAnswers: [],
-  audiocallShouldContinue: false,
   audiocallWrongAnswers: [],
   audiocallCorrectAnswers: [],
 };
@@ -24,15 +26,6 @@ const audiocallSlice = createSlice({
     setAudiocallPage: (state, action: PayloadAction<number>) => {
       state.audiocallPage = action.payload;
     },
-    setAudiocallCurrentWord: (state, action: PayloadAction<number>) => {
-      state.audiocallCurrentWord = action.payload;
-    },
-    setAudiocallAnswers: (state, action: PayloadAction<AudiocallAnswerInfo[]>) => {
-      state.audiocallAnswers = action.payload;
-    },
-    setAudiocallShouldContinue: (state, action: PayloadAction<boolean>) => {
-      state.audiocallShouldContinue = action.payload;
-    },
     setAudiocallWrongAnswers: (state, action: PayloadAction<Word[]>) => {
       state.audiocallWrongAnswers = action.payload;
     },
@@ -45,21 +38,12 @@ const audiocallSlice = createSlice({
 export const {
   setAudiocallGroup,
   setAudiocallPage,
-  setAudiocallCurrentWord,
-  setAudiocallAnswers,
-  setAudiocallShouldContinue,
   setAudiocallCorrectAnswers,
   setAudiocallWrongAnswers,
 } = audiocallSlice.actions;
 
 export const selectAudiocallGroup = (state: RootState): number => state.audiocall.audiocallGroup;
 export const selectAudiocallPage = (state: RootState): number => state.audiocall.audiocallPage;
-export const selectAudiocallCurrentWord = (state: RootState): number =>
-  state.audiocall.audiocallCurrentWord;
-export const selectAudiocallAnswers = (state: RootState): AudiocallAnswerInfo[] =>
-  state.audiocall.audiocallAnswers;
-export const selectAudiocallShouldContinue = (state: RootState): boolean =>
-  state.audiocall.audiocallShouldContinue;
 export const selectAudiocallWrongAnswers = (state: RootState): Word[] =>
   state.audiocall.audiocallWrongAnswers;
 export const selectAudiocallCorrectAnswers = (state: RootState): Word[] =>
